working on restricted pages

+++ studybud/base/views.py
@@ -80,7 +80,7 @@ def updateRoom(request,pk):
      room = Room.objects.get(id=pk)
      form = RoomForm(instance = room)
 
-     if request.user != room.user:
+     if request.user != room.host:
           return HttpResponse('your are not allowed')
 
      if request.method == 'POST':
@@ -95,6 +95,9 @@ def updateRoom(request,pk):
 @login_required(login_url="login")
 def deleteRoom(request, pk):
      room  = Room.objects.get(id=pk)
+     if request.user != room.host:
+          return HttpResponse('your are not allowed')
+
      if request.method == 'POST':
           room.delete()
           return redirect('home')
